Split SvgRenderer into focused helpers
